Updated DNNL tests to use pattern table mechanism
