+++ python/tvm/relay/op/contrib/dnnl.py
@@ -32,7 +32,9 @@ it is supported. For example:
 - The other way is to implement the function by themselves to
 check the attributes of the op and decide if it should be offloaded to DNNL.
 """
+from ... import expr as _expr
 from ... import op as reg
+from .register import register_pattern_table
 
 
 def _register_external_op_helper(op_name, supported=True):
@@ -63,3 +65,23 @@ _register_external_op_helper("nn.relu")
 _register_external_op_helper("add")
 _register_external_op_helper("subtract")
 _register_external_op_helper("multiply")
+
+
+def make_pattern(with_bias=True):
+    data = _expr.var("data")
+    weight = _expr.var("weight")
+    bias = _expr.var("bias")
+    conv = reg.nn.conv2d(data, weight)
+    if with_bias:
+        conv_out = reg.add(conv, bias)
+    else:
+        conv_out = conv
+    return reg.nn.relu(conv_out)
+
+
+@register_pattern_table("dnnl")
+def pattern_table():
+    conv2d_bias_relu_pat = ("dnnl.conv2d_bias_relu", make_pattern(with_bias=True))
+    conv2d_relu_pat = ("dnnl.conv2d_relu", make_pattern(with_bias=False))
+    dnnl_patterns = [conv2d_bias_relu_pat, conv2d_relu_pat]
+    return dnnl_patterns
